refactor links to Navigation.js
